script: Move remaining Document-related focus code to `dom/document/focus.rs`
This code was originally in `event_hander.rs` and this change removes
the remaining code to the appropriate file (which was created later).
